move sprite, but can't control it
Phaser's documentation (and the developers' writings) are all wrong. You
don't redraw a sprite at a new position with "sprite.x = x", it's
"sprite.body.x = x". sprite moves with game.physics.moveTo(), but it
moves in the wrong direction starting with the second call. So much more
work to be done...